MCU LCD Debugger: Step 1 of 3
LCD display will now show current ASM instruction and PC. Switches N17
and H18 control nClear and the CPU clock pause, respectively.

Additionally, support has been added for the quadrature encoder knob and
a single button, which currently only control the bank of 8 on-board
LEDs. Eventually they will control what is displayed on the LCD: between
current ASM instruction, registers, and DRAM.

A Xilinx ISE project has been added to the repo; it will stay up-to-date
while I continue to play around with my local project files on things I
don't want added to the repo. Also, this should clear up confusion with
the files in the "src" folder; not all of them are used, and some are
outdated.