MCU LCD Debugger: Step 2 of 3
Button D18 will switch between four display states, the last two of
which are unused and show the word "Default" (because it was easiest).
The first shows the current assembly instruction and program counter,
and the second shows the current values in the register memory. The
quadrature encoder knob can be used to adjust which registers are
displayed.

Note: the included Xilinx ISE project has NOT been updated to this
latest version. When the LCD debugger is complete, however, it will.